Editor Camera Controls with Lerp #33

+++ src/main/java/jade/GameObject.java
@@ -39,7 +39,7 @@ public class GameObject {
         }
     }
 
-    public void update(double deltaTime) {
+    public void update(float deltaTime) {
         for (Component c : components) {
              c.update(deltaTime);
         }
